move time reparam from abstract model to abstract multivariate model
